+++ src/Rhisis.World/Packets/TextMessagePackets.cs
@@ -52,16 +52,40 @@ namespace Rhisis.World.Packets
         /// </summary>
         /// <param name="player"></param>
         /// <param name="textId"></param>
-        /// <param name="stringParameter"></param>
-        public static void SendDefinedText(IPlayerEntity player, DefineText textId, params string[] stringParameter)
+        /// <param name="parameters"></param>
+        public static void SendDefinedText(IPlayerEntity player, DefineText textId, params object[] parameters)
         {
             using (var packet = new FFPacket())
             {
                 packet.StartNewMergedPacket(player.Id, SnapshotType.DEFINEDTEXT);
                 packet.Write((int)textId);
-                packet.Write(string.Join(" ", stringParameter));
+                packet.Write(string.Join(" ", parameters));
                 player.Connection.Send(packet);
             }
         }
+
+        /// <summary>
+        /// Sends a system message to the player.
+        /// </summary>
+        /// <param name="player">Player.</param>
+        /// <param name="text">System message text.</param>
+        public static void SendSnoop(IPlayerEntity player, string text)
+        {
+            using (var packet = new FFPacket())
+            {
+                packet.StartNewMergedPacket(0, SnapshotType.SNOOP);
+                packet.Write(text);
+
+                player.Connection.Send(packet);
+            }
+        }
+
+        /// <summary>
+        /// Sends a message to the player telling that the feature is not implemented yet.
+        /// </summary>
+        /// <param name="player">Player.</param>
+        /// <param name="feature">Feature name.</param>
+        public static void SendFeatureNotImplemented(IPlayerEntity player, string feature) 
+            => SendSnoop(player, $"Not implemented: {feature}");
     }
 }