+++ src/Rhisis.World/Systems/Battle/MeleeAttackArbiter.cs
@@ -45,10 +45,7 @@ namespace Rhisis.World.Systems.Battle
             
             if (this._attacker is IPlayerEntity player)
             {
-                Item rightWeapon = player.Inventory.GetItem(x => x.Slot == InventorySystem.RightWeaponSlot);
-
-                if (rightWeapon == null)
-                    rightWeapon = InventorySystem.Hand;
+                Item rightWeapon = player.Inventory.GetItem(x => x.Slot == InventorySystem.RightWeaponSlot) ?? InventorySystem.Hand;
 
                 // TODO: GetDamagePropertyFactor()
                 int weaponAttack = BattleHelper.GetWeaponAttackDamages(rightWeapon.Data.WeaponType, player);
@@ -100,7 +97,7 @@ namespace Rhisis.World.Systems.Battle
         {
             // TODO: if attacker mode == ONEKILL_MODE, return AF_GENERIC
 
-            int hitRate = 0;
+            int hitRate;
             int hitRating = this.GetHitRating(this._attacker);
             int escapeRating = this.GetEspaceRating(this._defender);
 
@@ -110,7 +107,7 @@ namespace Rhisis.World.Systems.Battle
                 hitRate = (int)(((hitRating * 1.5f) / (hitRating + escapeRating)) * 2.0f *
                           (this._attacker.Object.Level * 0.5f / (this._attacker.Object.Level + this._defender.Object.Level * 0.3f)) * 100.0f);
             }
-            else 
+            else
             {
                 // Player VS Player or Player VS Monster
                 hitRate = (int)(((hitRating * 1.6f) / (hitRating + escapeRating)) * 1.5f *
@@ -130,7 +127,7 @@ namespace Rhisis.World.Systems.Battle
         private int GetHitRating(ILivingEntity entity)
         {
             if (entity is IPlayerEntity player)
-                return player.Statistics.Dexterity; // TODO: add dex bonus
+                return player.Attributes[DefineAttributes.DEX]; // TODO: add dex bonus
             else if (entity is IMonsterEntity monster)
                 return monster.Data.HitRating;
 
@@ -145,7 +142,7 @@ namespace Rhisis.World.Systems.Battle
         public int GetEspaceRating(ILivingEntity entity)
         {
             if (entity is IPlayerEntity player)
-                return (int)(player.Statistics.Dexterity * 0.5f); // TODO: add dex bonus and DST_PARRY
+                return (int)(player.Attributes[DefineAttributes.DEX] * 0.5f); // TODO: add dex bonus and DST_PARRY
             else if (entity is IMonsterEntity monster)
                 return monster.Data.EscapeRating;
 
@@ -164,7 +161,7 @@ namespace Rhisis.World.Systems.Battle
                 return false;
 
             float criticalJobFactor = attacker is IPlayerEntity player ? player.PlayerData.JobData.Critical : 1f;
-            int criticalProbability = (int)((attacker.Statistics.Dexterity / 10) * criticalJobFactor);
+            int criticalProbability = (int)((attacker.Attributes[DefineAttributes.DEX] / 10) * criticalJobFactor);
             // TODO: add DST_CHR_CHANCECRITICAL to criticalProbability
 
             if (criticalProbability < 0)
@@ -261,7 +258,7 @@ namespace Rhisis.World.Systems.Battle
                 if (this._defender is IPlayerEntity player)
                     factor = player.PlayerData.JobData.DefenseFactor;
 
-                int stamina = this._defender.Statistics.Stamina;
+                int stamina = this._defender.Attributes[DefineAttributes.STA];
                 int level = this._defender.Object.Level;
                 int defense = (int)(((((level * 2) + (stamina / 2)) / 2.8f) - 4) + ((stamina - 14) * factor));
                 // TODO: add defense armor
@@ -299,10 +296,10 @@ namespace Rhisis.World.Systems.Battle
                     return 0.1f;
 
                 int defenderLevel = this._defender.Object.Level;
-                int defenderDexterity = this._defender.Statistics.Dexterity;
+                int defenderDexterity = this._defender.Attributes[DefineAttributes.DEX];
 
                 float blockProbabilityA = defenderLevel / ((defenderLevel + this._attacker.Object.Level) * 15.0f);
-                float blockProbabilityB = (defenderDexterity + this._attacker.Statistics.Dexterity + 2) * ((defenderDexterity - this._attacker.Statistics.Dexterity) / 800.0f);
+                float blockProbabilityB = (defenderDexterity + this._attacker.Attributes[DefineAttributes.DEX] + 2) * ((defenderDexterity - this._attacker.Attributes[DefineAttributes.DEX]) / 800.0f);
 
                 if (blockProbabilityB > 10.0f)
                     blockProbabilityB = 10.0f;